Divide EventsExpressService.js into separate elements
