+++ EventsExpress/ClientApp/src/actions/EditProfile/change-avatar.js
@@ -1,4 +1,4 @@
-import EventsExpressService from '../../services/EventsExpressService';
+import { UserService } from '../../services';
 import { setAlert } from '../alert';
 
 export const changeAvatar = {
@@ -8,7 +8,7 @@ export const changeAvatar = {
     UPDATE: "UPDATE_CHANGE_AVATAR"
 }
 
-const api_serv = new EventsExpressService();
+const api_serv = new UserService();
 
 export default function change_avatar(data) {
     return dispatch => {
