+++ EventsExpress/ClientApp/src/actions/add-category.js
@@ -1,4 +1,4 @@
-﻿import EventsExpressService from '../services/EventsExpressService';
+﻿import { CategoryService } from '../services';
 import get_categories from './category-list';
 
 
@@ -6,7 +6,7 @@ export const SET_CATEGORY_PENDING = "SET_CATEGORY_PENDING";
 export const SET_CATEGORY_SUCCESS = "SET_CATEGORY_SUCCESS";
 export const SET_CATEGORY_ERROR = "SET_CATEGORY_ERROR";
 
-const api_serv = new EventsExpressService();
+const api_serv = new CategoryService();
 
 
 export function add_category(data) {
