+++ EventsExpress/ClientApp/src/actions/add-copy-event.js
@@ -1,4 +1,4 @@
-import EventsExpressService from '../services/EventsExpressService';
+import { EventService } from '../services';
 import { setAlert } from './alert';
 import {createBrowserHistory} from 'history';
 
@@ -7,7 +7,7 @@ export const SET_COPY_EVENT_PENDING = "SET_COPY_EVENT_PENDING";
 export const SET_COPY_EVENT_ERROR = "SET_COPY_EVENT_ERROR";
 export const EVENT_COPY_WAS_CREATED = "EVENT_COPY_WAS_CREATED";
 
-const api_serv = new EventsExpressService();
+const api_serv = new EventService();
 const history = createBrowserHistory({forceRefresh:true});
 
 export default function add_copy_event(eventId) {
