+++ EventsExpress/ClientApp/src/actions/add-event.js
@@ -1,4 +1,4 @@
-import EventsExpressService from '../services/EventsExpressService';
+import { EventService } from '../services';
 import get_event from './event-item-view';
 
 export const SET_EVENT_SUCCESS = "SET_EVENT_SUCCESS";
@@ -6,7 +6,7 @@ export const SET_EVENT_PENDING = "SET_EVENT_PENDING";
 export const SET_EVENT_ERROR = "SET_EVENT_ERROR";
 export const EVENT_WAS_CREATED = "EVENT_WAS_CREATED";
 
-const api_serv = new EventsExpressService();
+const api_serv = new EventService();
 
 export default function add_event(data) {
