+++ EventsExpress/ClientApp/src/actions/delete-category.js
@@ -1,4 +1,4 @@
-﻿import EventsExpressService from '../services/EventsExpressService';
+﻿import { CategoryService } from '../services';
 import get_categories from './category-list';
 
 export const SET_CATEGORY_DELETE_PENDING = "SET_CATEGORY_DELETE_PENDING";
@@ -6,11 +6,9 @@ export const SET_CATEGORY_DELETE_SUCCESS = "SET_CATEGORY_DELETE_SUCCESS";
 export const SET_CATEGORY_DELETE_ERROR = "SET_CATEGORY_DELETE_ERROR";
 export const SET_CATEGORY_EDITED = "SET_CATEGORY_EDITED";
 
-
-
+const api_serv = new CategoryService();
 
 export function delete_category(data) {
-    const api_serv = new EventsExpressService();
 
     return dispatch => {
         dispatch(setCategoryDeletePending(true));
