+++ EventsExpress/ClientApp/src/actions/edit-event-from-parent.js
@@ -1,4 +1,4 @@
-import EventsExpressService from '../services/EventsExpressService';
+import { EventService } from '../services';
 import { setAlert } from './alert';
 import {createBrowserHistory} from 'history';
 
@@ -7,7 +7,7 @@ export const SET_EVENT_FROM_PARENT_PENDING = "SET_EVENT_FROM_PARENT_PENDING";
 export const SET_EVENT_FROM_PARENT_ERROR = "SET_EVENT_FROM_PARENT_ERROR";
 export const EVENT_FROM_PARENT_WAS_CREATED = "EVENT_FROM_PARENT_WAS_CREATED";
 
-const api_serv = new EventsExpressService();
+const api_serv = new EventService();
 const history = createBrowserHistory({forceRefresh:true});
 
 export default function edit_event_from_parent(data) {
