+++ EventsExpress/ClientApp/src/actions/inventory-list.js
@@ -1,10 +1,10 @@
-import EventsExpressService from '../services/EventsExpressService';
+import { InventoryService } from '../services';
 
 export const SET_INVENTORY_PENDING = "SET_INVENTORY_PENDING";
 export const GET_INVENTORY_SUCCESS = "GET_INVENTORY_SUCCESS";
 export const SET_INVENTORY_ERROR = "SET_INVENTORY_ERROR";
 
-const api_serv = new EventsExpressService();
+const api_serv = new InventoryService();
 
 export function get_inventories_by_event_id(eventId) {
     return dispatch => {
