+++ EventsExpress/ClientApp/src/actions/unitsOfMeasuring.js
@@ -1,8 +1,8 @@
-import EventsExpressService from '../services/EventsExpressService';
+import { InventoryService } from '../services';
 
 export const GET_UNITS_OF_MEASURING = 'GET_UNITS_OF_MEASURING';
 
-const api_serv = new EventsExpressService();
+const api_serv = new InventoryService();
 
 export default function get_unitsOfMeasuring() {
     return dispatch => {
