+++ EventsExpress/ClientApp/src/services/index.js
@@ -1,0 +1,23 @@
+import AuthenticationService from './AuthenticationService'
+import CategoryService from './CategoryService'
+import ChatService from './ChatService'
+import CommentService from './CommentService'
+import EventService from './EventService'
+import EventScheduleService from './EventScheduleService'
+import InventoryService from './InventoryService'
+import LocationService from './LocationService'
+import RoleService from './RoleService'
+import UserService from './UserService'
+
+export {
+    AuthenticationService,
+    CategoryService,
+    ChatService,
+    CommentService,
+    EventService, 
+    EventScheduleService,
+    InventoryService,
+    LocationService,
+    RoleService,
+    UserService
+};
