+++ EventsExpress/Controllers/EventController.cs
@@ -37,9 +37,9 @@ namespace EventsExpress.Controllers
         /// <param name="model">Required.</param>
         /// <response code="200">Create event proces success.</response>
         /// <response code="400">If Create process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("[action]/{eventId:Guid}")]
         [UserAccessTypeFilter]
-        public async Task<IActionResult> EditEventFromParent(Guid eventId, [FromForm] EventEditViewModel model)
+        public async Task<IActionResult> CreateNextFromParentWithEdit(Guid eventId, [FromForm] EventEditViewModel model)
         {
             if (!ModelState.IsValid)
             {
@@ -57,8 +57,8 @@ namespace EventsExpress.Controllers
         /// <param name="eventId">Required.</param>
         /// <response code="200">Create event proces success.</response>
         /// <response code="400">If Create process failed.</response>
-        [HttpPost("[action]")]
-        public async Task<IActionResult> CreateEventFromParent(Guid eventId)
+        [HttpPost("[action]/{eventId:Guid}")]
+        public async Task<IActionResult> CreateNextFromParent(Guid eventId)
         {
             if (!ModelState.IsValid)
             {
@@ -95,7 +95,7 @@ namespace EventsExpress.Controllers
         /// <param name="model">Required.</param>
         /// <response code="200">Edit event proces success.</response>
         /// <response code="400">If Edit process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [UserAccessTypeFilter]
         public async Task<IActionResult> Edit(Guid eventId, [FromForm] EventEditViewModel model)
         {
@@ -112,13 +112,13 @@ namespace EventsExpress.Controllers
         /// <summary>
         /// This method have to return event.
         /// </summary>
-        /// <param name="id">Required.</param>
+        /// <param name="eventId">Required.</param>
         /// <returns>Event.</returns>
         /// <response code="200">Return UserInfo model.</response>
         [AllowAnonymous]
-        [HttpGet("[action]")]
-        public IActionResult Get(Guid id) =>
-            Ok(_mapper.Map<EventViewModel>(_eventService.EventById(id)));
+        [HttpGet("{eventId:Guid}")]
+        public IActionResult Get(Guid eventId) =>
+            Ok(_mapper.Map<EventViewModel>(_eventService.EventById(eventId)));
 
         /// <summary>
         /// This method have to return all events.
@@ -169,11 +169,11 @@ namespace EventsExpress.Controllers
         /// <summary>
         /// This method have to add user to category.
         /// </summary>
-        /// <param name="userId">Required.</param>
         /// <param name="eventId">EventId.</param>
+        /// <param name="userId">Required.</param>
         /// <response code="200">Adding user from event proces success.</response>
         /// <response code="400">If adding user from event process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         public async Task<IActionResult> AddUserToEvent(Guid eventId, Guid userId)
         {
             await _eventService.AddUserToEvent(userId, eventId);
@@ -184,11 +184,11 @@ namespace EventsExpress.Controllers
         /// <summary>
         /// This method have to approved user on event.
         /// </summary>
-        /// <param name="userId">Required.</param>
         /// <param name="eventId">EventId.</param>
+        /// <param name="userId">Required.</param>
         /// <response code="200">Approving user from event process success.</response>
         /// <response code="400">If aproving user from event process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [UserAccessTypeFilter]
         public async Task<ActionResult> ApproveVisitor(Guid eventId, Guid userId)
         {
@@ -204,7 +204,7 @@ namespace EventsExpress.Controllers
         /// <param name="eventId">EventId.</param>
         /// <response code="200">Denying user from event process success.</response>
         /// <response code="400">If denying user from event process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [UserAccessTypeFilter]
         public async Task<ActionResult> DenyVisitor(Guid eventId, Guid userId)
         {
@@ -216,11 +216,11 @@ namespace EventsExpress.Controllers
         /// <summary>
         /// This method have to add user to category.
         /// </summary>
-        /// <param name="userId">Required.</param>
         /// <param name="eventId">EventId.</param>
+        /// <param name="userId">Required.</param>
         /// <response code="200">Delete  user from event proces success.</response>
         /// <response code="400">If deleting user from event process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         public async Task<IActionResult> DeleteUserFromEvent(Guid eventId, Guid userId)
         {
             await _eventService.DeleteUserFromEvent(userId, eventId);
@@ -235,7 +235,7 @@ namespace EventsExpress.Controllers
         /// <response code="200">Block is succesful.</response>
         /// <response code="302">If user isn't admin.</response>
         /// <response code="400">Block process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Block(Guid eventId)
         {
@@ -250,7 +250,7 @@ namespace EventsExpress.Controllers
         /// <param name="eventId">Required.</param>
         /// <response code="200">Unblock is succesful.</response>
         /// <response code="400">Unblock process is failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Unblock(Guid eventId)
         {
@@ -284,7 +284,7 @@ namespace EventsExpress.Controllers
         /// <param name="eventId">Required (type: Guid).</param>
         /// <response code="200">Getting is successful.</response>
         /// <response code="400">Getting is failed.</response>
-        [HttpGet("[action]")]
+        [HttpGet("{eventId:Guid}/[action]")]
         public IActionResult GetCurrentRate(Guid eventId)
         {
             if (!_eventService.Exists(eventId))
@@ -304,7 +304,7 @@ namespace EventsExpress.Controllers
         /// <returns>RateOfEvent.</returns>
         /// <response code="200">Getting is successful.</response>
         /// <response code="400">Getting is failed.</response>
-        [HttpGet("[action]")]
+        [HttpGet("{eventId:Guid}/[action]")]
         public IActionResult GetAverageRate(Guid eventId)
         {
             if (!_eventService.Exists(eventId))
