+++ EventsExpress/Controllers/EventScheduleController.cs
@@ -58,7 +58,7 @@ namespace EventsExpress.Controllers
         /// <param name="model">Required.</param>
         /// <response code="200">Edit/Create event proces success.</response>
         /// <response code="400">If Edit/Create process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [UserAccessTypeFilter]
         public async Task<IActionResult> Edit(Guid eventId, [FromForm] EventScheduleViewModel model)
         {
@@ -73,7 +73,7 @@ namespace EventsExpress.Controllers
         /// <param name="eventId">Required.</param>
         /// <response code="200">Cancel All Events proces success.</response>
         /// <response code="400">Cancel All Events process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [UserAccessTypeFilter]
         public async Task<IActionResult> CancelAllEvents(Guid eventId)
         {
@@ -93,7 +93,7 @@ namespace EventsExpress.Controllers
         /// <param name="eventId">Required.</param>
         /// <response code="200">Cancel Next Event event proces success.</response>
         /// <response code="400">Cancel Next Event process failed.</response>
-        [HttpPost("[action]")]
+        [HttpPost("{eventId:Guid}/[action]")]
         [UserAccessTypeFilter]
         public async Task<IActionResult> CancelNextEvent(Guid eventId)
         {
@@ -114,7 +114,7 @@ namespace EventsExpress.Controllers
         /// <returns>Event.</returns>
         /// <response code="200">Return UserInfo model.</response>
         [AllowAnonymous]
-        [HttpGet("[action]")]
+        [HttpGet("{id:Guid}")]
         public IActionResult Get(Guid id) =>
             Ok(_mapper.Map<EventScheduleViewModel>(_eventScheduleService.EventScheduleById(id)));
     }
